mcp-server-rag-docs/pyproject.toml
[project]
name = "mcp-server-rag-docs"
version = "0.1.0"
description = "MCP RAG docs server with in-memory FAISS index"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.4.1",
    "langchain-text-splitters>=0.3.0",
    "faiss-cpu>=1.8.0",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
]
